requirements.txt
annoy
arrow
chainlit
cohere
datasets
fastapi
gensim
huggingface-hub
icecream
instructor
ipykernel
ipython
Jinja2
jupyter
jupyterlab
llama-index
matplotlib
mistralai
mplcursors
networkx
nltk
notebook
numpy
openai
pandas
panel
Pillow
pip
pydantic
python-dateutil
python-dotenv
redlines
regex
requests
rich
safetensors
scikit-learn
scipy
sentence-transformers
sentencepiece
setuptools
six
SQLAlchemy
sympy
tiktoken
tokenizers
torch
torchaudio
torchmetrics
torchvision
tornado
tqdm
transformers
typer
urllib3
uvicorn
weaviate-client
wheel
